realloc

+++ 0x0C-more_malloc_free/0-malloc_checked.c
@@ -9,7 +9,7 @@
  */
 void *malloc_checked(unsigned int b)
 {
-	b = malloc(sizeof(b));
+	b = (unsigned int)malloc(sizeof(b));
 
 	if (b == NULL)
 	{
